Fix bug with re-applying same patch twice
there was condition when regexp can capture redundant leading space
which lead to applying patch twice. Patch was applied from cached stream
object and this stream object wasn't reset back to beginning of the
stream after reusing, which confused git and was considered as an empty
patch, which created condition when patch failed to apply, tagged,
aplied with an empty patch and untagged again.